Find a list of package sources
