+++ evaluation/package.go
@@ -6,6 +6,7 @@ import "github.com/Zac-Garby/pluto/ast"
 type Package struct {
 	Context *Context // the context containing the package's functions
 	Used    bool     // whether 'use <this package>' has been called
+	Sources []string // unglobbed source files
 
 	/* metadata read from 'collections.yaml' */
 	Meta struct {
